+++ step-by-step/step_crawler/functions_file.py
@@ -1,11 +1,25 @@
 import asyncio
 import time
 
+
 def range_(stop):
     return [i for i in range(stop)]
 
+
 def print_(word):
     return word
+
+
+def espere(segs):
+    time.sleep(segs)
+
+
+async def wait_page(page):
+    jsWait = "document.readyState === 'complete' || \
+              document.readyState === 'iteractive'"
+    while not (await page.evaluate(jsWait)):
+        await page.waitFor(1)
+
 
 async def clique(page, xpath):
     await page.waitForXPath(xpath)
@@ -38,6 +52,7 @@ async def elementos_nesse_xpath(page, xpath):
 
 def espere(segs):
 	time.sleep(segs)
+
 
 async def for_clicavel(page, xpath):
     try:
