+++ step-by-step/tests/code_generator_test.py
@@ -1,19 +1,17 @@
-import sys
-sys.path.append("../code")
 import unittest
-import code_generator as cg
 import json
-import atomizer as atom
+
+from step_crawler import code_generator as cg
+from step_crawler import atomizer as atom
 
 
 class TestExtractInfo(unittest.TestCase):
     def test_dict_to_arguments(self):
-
-        result = cg.dict_to_arguments({'a':'a', 'b':'b', 'c':'c'})
+        result = cg.dict_to_arguments({'a': 'a', 'b': 'b', 'c': 'c'})
         expected_result = "a = a, b = b, c = c"
         self.assertEqual(expected_result, result)
 
-        result = cg.dict_to_arguments({'a':'a'})
+        result = cg.dict_to_arguments({'a': 'a'})
         expected_result = "a = a"
         self.assertEqual(expected_result, result)
 
@@ -22,12 +20,11 @@ class TestExtractInfo(unittest.TestCase):
         self.assertEqual(expected_result, result)
 
     def test_generate_call(self):
-
-        result = cg.generate_call("print", {'a':'a', 'b':'b', 'c':'c'})
+        result = cg.generate_call("print", {'a': 'a', 'b': 'b', 'c': 'c'})
         expected_result = "print(a = a, b = b, c = c)"
         self.assertEqual(expected_result, result)
 
-        result = cg.generate_call("print", {'a':'a'})
+        result = cg.generate_call("print", {'a': 'a'})
         expected_result = "print(a = a)"
         self.assertEqual(expected_result, result)
 
@@ -35,11 +32,13 @@ class TestExtractInfo(unittest.TestCase):
         expected_result = "print()"
         self.assertEqual(expected_result, result)
 
-        result = cg.generate_call("print", {'a':'a', 'b':'b', 'c':'c'}, True)
-        expected_result = "await print(**missing_arguments, a = a, b = b, c = c)"
+        result = cg.generate_call("print", {'a': 'a', 'b': 'b', 'c': 'c'},
+                                  True)
+        expected_result = "await print(**missing_arguments, " \
+                          "a = a, b = b, c = c)"
         self.assertEqual(expected_result, result)
 
-        result = cg.generate_call("print", {'a':'a'}, True)
+        result = cg.generate_call("print", {'a': 'a'}, True)
         expected_result = "await print(**missing_arguments, a = a)"
         self.assertEqual(expected_result, result)
 
@@ -49,29 +48,34 @@ class TestExtractInfo(unittest.TestCase):
 
     def test_generate_head(self):
         result = cg.generate_head(json)
-        expected_result = "import sys \n" + "sys.path.append(\"code\")\n"
+        expected_result = "import step_crawler\n"
         expected_result = expected_result + "from " + "json" + " import *\n\n"
-        expected_result = expected_result + "async def execute_steps(**missing_arguments):\n"
+        expected_result = expected_result + "async def " \
+                                            "execute_steps(" \
+                                            "**missing_arguments):\n"
         self.assertEqual(expected_result, result)
 
     def test_generate_body(self):
-        with open("recipe_examples.json") as file:
+        with open("../recipe_examples.json") as file:
             recipe_examples = json.load(file)
-        ff = __import__("functions_file")
+        ff = __import__("step_crawler").functions_file
 
-        result = cg.generate_body(recipe_examples['unbreakable_between_breakable']['recipe'], ff)
+        result = cg.generate_body(
+            recipe_examples['unbreakable_between_breakable']['recipe'], ff)
         expected_result = "    for i in range_(stop = 2):\n"
-        expected_result +="        for j in range_(stop = 2):\n" 
-        expected_result +="            for k in range_(stop = 2):\n" 
-        expected_result +="                print_(word = \"teste\")\n" 
+        expected_result += "        for j in range_(stop = 2):\n"
+        expected_result += "            for k in range_(stop = 2):\n"
+        expected_result += "                print_(word = \"teste\")\n"
         self.assertEqual(expected_result, result)
 
-        result = cg.generate_body(atom.extend(recipe_examples['unbreakable_between_breakable']['recipe'])[0], ff)
+        result = cg.generate_body(atom.extend(
+            recipe_examples['unbreakable_between_breakable']['recipe'])[0], ff)
         expected_result = "    i = 0\n"
-        expected_result +="    for j in range_(stop = 2):\n" 
-        expected_result +="        k = 0\n" 
-        expected_result +="        print_(word = \"teste\")\n" 
-        self.assertEqual(expected_result, result)        
+        expected_result += "    for j in range_(stop = 2):\n"
+        expected_result += "        k = 0\n"
+        expected_result += "        print_(word = \"teste\")\n"
+        self.assertEqual(expected_result, result)
+
 
 if __name__ == '__main__':
     unittest.main()
