+++ step-by-step/tests/crawler-tests/diario-oficial-uniao/main.py
@@ -1,16 +1,10 @@
 import json
-import sys
 
-sys.path.append('../../../code')
-import code_generator as code_g
-import atomizer as atom
-from functions_file import *
-import functions_file
-
-import time
 from pyppeteer import launch
 
-
+from step_crawler import code_generator as code_g
+from step_crawler import functions_file
+from step_crawler.functions_file import *
 
 
 async def main():
@@ -19,7 +13,7 @@ async def main():
     await page.goto('http://www.in.gov.br/web/guest/inicio')
 
     steps = __import__('diario_oficial_uniao_steps')
-    await steps.execute_steps(page = page)
+    await steps.execute_steps(page=page)
 
     await browser.close()
     return
@@ -33,11 +27,3 @@ with open('diario_oficial_uniao_steps.py', 'w+') as file:
     file.write(code_g.generate_code(recipe, functions_file))
 
 asyncio.get_event_loop().run_until_complete(main())
-
-
-
-
-
-
-
-
